python api client logging and documentation

+++ examples/python/job-cli.py
@@ -1,16 +1,8 @@
 import argparse
-from datetime import datetime
-import logging
 import os
 import re
 import requests
 import time
-
-logging.basicConfig()
-logging.getLogger().setLevel(logging.DEBUG)
-requests_log = logging.getLogger("requests.packages.urllib3")
-requests_log.setLevel(logging.DEBUG)
-requests_log.propagate = True
 
 
 class ApiError(Exception):
@@ -64,11 +56,18 @@ class Action:
         if self.token is None:
             return self.retrieve_token()
 
-        expires_at = self.token_generated + self.token['expires'] - Action.EXPIRATION_BUFFER
+        expires_at = self.token_generated + self.token['expires_in'] - Action.EXPIRATION_BUFFER
         if expires_at < time.time():
+            print('Refreshing JWT Token')
             return self.retrieve_token()
 
         return self.token
+
+    @staticmethod
+    def print_headers(response):
+        for header in response.headers.items():
+            print('%s: %s' % header)
+        print()
 
 
 class StartJob(Action):
@@ -92,6 +91,8 @@ class StartJob(Action):
 
     def __call__(self, *args, **kwargs):
 
+        print("Attempting to start a new job\n")
+
         token = self.retrieve_token()
 
         url = self.start_url()
@@ -105,16 +106,22 @@ class StartJob(Action):
             "_type": "ExplanationOfBenefit"
         })
 
+        self.print_headers(response)
+
         status = response.status_code
         if status != 200 and status != 202:
             raise ApiError("Could not start job at url %s. Response %d. Message %s"
                            % (url, response.status_code, response.json()))
+
+        print("Successfully started a job")
 
         status_url = response.headers['content-location']
         job_id = re.findall('Job/(.*)/\\$status', status_url)[0]
 
         with open(self.job_id_file, 'w+') as job_file:
             job_file.write(job_id)
+
+        print("Saved job id to %s\n" % self.job_id_file)
 
 
 class MonitorJob(Action):
@@ -129,7 +136,6 @@ class MonitorJob(Action):
         super().__init__(idp_url, api_url, auth)
         self.job_id_path = job_id_path
         self.response_path = response_path
-        self.url = self.get_monitor_url()
 
     def get_monitor_url(self):
 
@@ -138,29 +144,44 @@ class MonitorJob(Action):
 
             return "%s/Job/%s/$status" % (self.api_url, job_id)
 
-    def check_status(self):
+    def check_status(self, url):
         token = self.get_or_refresh_token()
-        return requests.get(self.url, headers={
+        response = requests.get(url, headers={
             "Accept": "application/json",
             "Prefer": "respond-async",
             "Authorization": "Bearer %s" % token['access_token']
         })
 
+        self.print_headers(response)
+
+        return response
+
     def __call__(self, *args, **kwargs):
 
-        response = self.check_status()
+        print("Attempting to monitor job's status\n")
+
+        # Sleep so API doesn't throw querying too often error
+        time.sleep(5)
+
+        url = self.get_monitor_url()
+
+        response = self.check_status(url)
         while response.status_code != 200:
-            if response == 202:
+            if response.status_code == 202:
                 time.sleep(60)
-                response = self.check_status()
+                response = self.check_status(url)
             else:
                 raise ApiError("Could not retrieve bearer token from %s. Response %d. Message %s"
-                               % (self.url, response.status_code, response.json()))
+                               % (url, response.status_code, response.json()))
 
+        print("Job has finished, saving files to download to %s" % self.response_path)
         with open(self.response_path, 'w+') as response_file:
             output = response.json()['output']
             for file in output:
                 response_file.write(file['url'])
+                response_file.write('\n')
+
+        print("Finished monitoring job\n")
 
 
 class DownloadResults(Action):
@@ -181,31 +202,39 @@ class DownloadResults(Action):
         local_filename = url.split('/')[-1]
         local_path = self.directory + os.sep + local_filename
 
-        # NOTE the stream=True parameter below
-        with requests.get(url, headers={
-            "Accept": "application/fhir+ndjson",
-            "Authorization": "Bearer %s" % token['access_token']
-        }, stream=True) as r:
+        if not os.path.exists(local_path):
+            print("Downloading %s to %s" % (local_filename, self.directory))
 
-            # Raise error on failure of the call
-            r.raise_for_status()
+            # NOTE the stream=True parameter below
+            with requests.get(url, headers={
+                "Accept": "application/fhir+ndjson",
+                "Authorization": "Bearer %s" % token['access_token']
+            }, stream=True) as r:
 
-            # Write out to a file
-            with open(local_path, 'wb') as f:
-                for chunk in r.iter_content(chunk_size=8192):
-                    # If you have chunk encoded response uncomment if
-                    # and set chunk_size parameter to None.
-                    #if chunk:
-                    f.write(chunk)
-        return local_filename
+                # Raise error on failure of the call
+                r.raise_for_status()
+
+                # Write out to a file
+                with open(local_path, 'wb') as f:
+                    for chunk in r.iter_content(chunk_size=8192):
+                        # If you have chunk encoded response uncomment if
+                        # and set chunk_size parameter to None.
+                        #if chunk:
+                        f.write(chunk)
+        else:
+            print("WARNING %s already exists will not overwrite by downloading export file\n" % local_path)
 
     def __call__(self, *args, **kwargs):
 
+        print("Attempting to download results of job\n")
+
         with open(self.response_path, 'r') as response_file:
-            urls = response_file.readlines()
+            urls = response_file.read().splitlines()
 
             for url in urls:
                 self.download(url)
+
+        print("Finished downloading results")
 
 
 def get_env(args):
@@ -267,10 +296,9 @@ try:
     else:
         tasks.append(StartJob(idp_url, api_url, args.auth, job_id_path))
         tasks.append(MonitorJob(idp_url, api_url, args.auth, job_id_path, completion_id_path))
-        tasks.append(DownloadResults(idp_url, api_url, args.auth, completion_id_path))
+        tasks.append(DownloadResults(idp_url, api_url, args.auth, args.directory, completion_id_path))
 
     for task in tasks:
-        print(task.__class__)
         task()
 
 except ValueError as ve:
